+++ docs/ChangeLog.md
@@ -114,6 +114,7 @@
 - auto-complete: add `bleopt complete_auto_complete_opts` (motivated by DUOLabs333) `#D1901` `#D1911` 1478a04 xxxxxxx
 - menu-complete: add `bleopt complete_menu_complete_opts` (requested by DUOLabs333) `#D1911` xxxxxxx
 - edit (`magic-space`): support `bleopt edit_magic_expand=...:alias` (requested by telometto) `#D1912` xxxxxxx
+  - auto-complete: cancel auto-complete for `magic-space` `#D1913` xxxxxxx
 
 ## Changes